Update examples of BTCChina.

+++ xchange-examples/src/main/java/com/xeiam/xchange/examples/btcchina/trade/BTCChinaGetOrderDemo.java
@@ -19,7 +19,7 @@ public class BTCChinaGetOrderDemo {
 
   public static void main(String[] args) throws IOException {
 
-    final long orderId = Long.parseLong(args[0]);
+    final int orderId = Integer.parseInt(args[0]);
     final String market = args.length > 1 ? args[1] : null;
 
     BTCChinaGetOrderResponse response;
